Extract grunt task
